+++ Task/RemoveFileTask.php
@@ -12,7 +12,7 @@ namespace CleverAge\FlysystemProcessBundle\Task;
 
 use CleverAge\ProcessBundle\Model\AbstractConfigurableTask;
 use CleverAge\ProcessBundle\Model\ProcessState;
-use League\Flysystem\MountManager;
+use League\Flysystem\FilesystemException;
 use Psr\Log\LoggerInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
@@ -23,46 +23,31 @@ class RemoveFileTask extends AbstractConfigurableTask
 {
     use FilesystemOptionTrait;
 
-    /** @var  MountManager */
-    protected $mountManager;
-
-    /** @var LoggerInterface */
-    protected $logger;
-
-    /**
-     * RemoveFileTask constructor.
-     *
-     * @param MountManager    $mountManager
-     * @param LoggerInterface $logger
-     */
-    public function __construct(MountManager $mountManager, LoggerInterface $logger)
+    public function __construct(protected LoggerInterface $logger)
     {
-        $this->mountManager = $mountManager;
-        $this->logger = $logger;
     }
 
-    protected function configureOptions(OptionsResolver $resolver)
+    protected function configureOptions(OptionsResolver $resolver): void
     {
         $this->configureFilesystemOption($resolver, 'filesystem');
     }
 
-    public function execute(ProcessState $state)
+    public function execute(ProcessState $state): void
     {
         $filesystem = $this->getFilesystem($state, 'filesystem');
         $filePath = $state->getInput();
 
-        $success = $filesystem->delete($filePath);
+        try {
+            $filesystem->delete($filePath);
+            $result = true;
+        } catch (FilesystemException) {
+            $result = false;
+        }
 
-        if ($success) {
+        if ($result) {
             $this->logger->info('Deleted input file', ['file' => $filePath]);
         } else {
             $this->logger->warning('Failed to deleted input file', ['file' => $filePath]);
         }
     }
-
-    protected function getMountManager(): MountManager
-    {
-        return $this->mountManager;
-    }
-
 }