+++ CHANGELOG.md
@@ -3,6 +3,9 @@ v2.0
 
 ## BC breaks
 
+* [#5](https://github.com/cleverage/flysystem-process-bundle/issues/5) Replace "oneup/flysystem-bundle": ">1.0,<4.0" by "league/flysystem-bundle": "^3.0"
+* [#5](https://github.com/cleverage/flysystem-process-bundle/issues/5) Update Tasks for "league/flysystem-bundle": "^3.0"
+
 ### Changes
 
 ### Fixes